// src/shared/api/@generated/models/recommended-unit-response.ts
/* tslint:disable */
 
/**
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface RecommendedUnitResponse {
    'unitId'?: number;
    'unitTitle'?: string;
    'chapterId'?: number;
    'chapterTitle'?: string;
}