// src/shared/api/@generated/models/weekly-learning-record-response.ts
/* tslint:disable */
 
/**
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface WeeklyLearningRecordResponse {
    'MONDAY'?: boolean;
    'TUESDAY'?: boolean;
    'WEDNESDAY'?: boolean;
    'THURSDAY'?: boolean;
    'FRIDAY'?: boolean;
    'SATURDAY'?: boolean;
    'SUNDAY'?: boolean;
}